+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed AoE with Hit Locations and now show more hit location info in to-hit dialog. [#1768](https://github.com/dmdorman/hero6e-foundryvtt/issues/1768)
 - Improved default FoundryVTT appendNumber. Adding single tokens to a scene that appends a number to the token now ensure duplicate token names are not created. Does not change the default FoundryVTT behavior for adding multiple tokens during a paste operation.
 - Combat maneuvers which don't use STR now use 1 END.
+- Automatons now have perception and maneuvers again. [#1756](https://github.com/dmdorman/hero6e-foundryvtt/issues/1756)
 
 ## Version 4.0.14 20250119